% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{bernoulli_process}
\alias{bernoulli_process}
\title{Bernoulli process}
\usage{
bernoulli_process(variable, from, to, rate)
}
\arguments{
\item{variable}{a categorical variable}

\item{from}{a string representing the source category}

\item{to}{a string representing the destination category}

\item{rate}{the rate to move individuals between categories}
}
\description{
Updates categorical variable from the \code{from} state to the \code{to}
state at the rate \code{rate}
}
