% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitset.R
\name{Bitset}
\alias{Bitset}
\title{Class: Bitset
An R wrapper for IterableBitset}
\description{
Class: Bitset
An R wrapper for IterableBitset

Class: Bitset
An R wrapper for IterableBitset
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Bitset$new()}}
\item \href{#method-insert}{\code{Bitset$insert()}}
\item \href{#method-remove}{\code{Bitset$remove()}}
\item \href{#method-size}{\code{Bitset$size()}}
\item \href{#method-or}{\code{Bitset$or()}}
\item \href{#method-and}{\code{Bitset$and()}}
\item \href{#method-not}{\code{Bitset$not()}}
\item \href{#method-sample}{\code{Bitset$sample()}}
\item \href{#method-copy}{\code{Bitset$copy()}}
\item \href{#method-to_vector}{\code{Bitset$to_vector()}}
\item \href{#method-clone}{\code{Bitset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a bitset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$new(size, from = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{the size of the bitset}

\item{\code{from}}{pointer to an existing IterableBitset to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert"></a>}}
\if{latex}{\out{\hypertarget{method-insert}{}}}
\subsection{Method \code{insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$insert(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$remove(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-or"></a>}}
\if{latex}{\out{\hypertarget{method-or}{}}}
\subsection{Method \code{or()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$or(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-and"></a>}}
\if{latex}{\out{\hypertarget{method-and}{}}}
\subsection{Method \code{and()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$and(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-not"></a>}}
\if{latex}{\out{\hypertarget{method-not}{}}}
\subsection{Method \code{not()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$not()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$sample(rate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_vector"></a>}}
\if{latex}{\out{\hypertarget{method-to_vector}{}}}
\subsection{Method \code{to_vector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$to_vector()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bitset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
