% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitset.R
\name{bitset_count_and}
\alias{bitset_count_and}
\title{Count bitset and}
\usage{
bitset_count_and(a, b)
}
\arguments{
\item{a}{a \code{\link{Bitset}}}

\item{b}{another \code{\link{Bitset}}}
}
\description{
This non-modifying function returns the number of intersecting
elements between two bitsets \code{a} and \code{b}. This should be faster than
writing \code{a$copy()$and(b)$size()} as it avoids the memory allocations of $copy().
}
