% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{checkpoint_state}
\alias{checkpoint_state}
\title{Save the simulation state}
\usage{
checkpoint_state(timesteps, variables, events)
}
\arguments{
\item{variables}{the list of Variables}

\item{events}{the list of Events}
}
\description{
Save the simulation state in an R object, allowing it to be
resumed later using \code{\link[individual]{restore_state}}.
}
