% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{Constant}
\alias{Constant}
\title{Class: Constant
Represents a constant for an individual in our simulation}
\description{
Class: Constant
Represents a constant for an individual in our simulation

Class: Constant
Represents a constant for an individual in our simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Constant$new()}}
\item \href{#method-print}{\code{Constant$print()}}
\item \href{#method-clone}{\code{Constant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new State
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$new(name, initialiser)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{is a unique idetifier which is used in the output}

\item{\code{initialiser}}{a function used to initialise the constant at the start
of the simulation. The initialiser function takes the population size as
its only argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
