% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_state}
\alias{create_state}
\title{Create a simulation state}
\usage{
create_state(individuals)
}
\arguments{
\item{individuals}{a list of individual objects}
}
\description{
Create a simulation state
}
