% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DataClass}
\alias{DataClass}
\title{Creates a DataClass design pattern
readonly class for collecting related fields
implementation inspired by https://github.com/r-lib/R6/issues/186}
\usage{
DataClass(classname, fields, initialize, print_fields = NULL)
}
\arguments{
\item{classname}{is the name of the class}

\item{fields}{is a sequence of readonly fields to create}

\item{initialize}{is the initialisation funtion}

\item{print_fields}{are the fields to print}
}
\description{
Creates a DataClass design pattern
readonly class for collecting related fields
implementation inspired by https://github.com/r-lib/R6/issues/186
}
