% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{Event}
\alias{Event}
\title{Class: Event
Describes an event in the simulation}
\description{
Class: Event
Describes an event in the simulation

Class: Event
Describes an event in the simulation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name, }}{the unique name of the event}

\item{\code{listeners, }}{the listener functions to be executed when the event is
triggered}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name, }}{the unique name of the event}

\item{\code{listeners, }}{the listener functions to be executed when the event is
triggered}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Event$new()}}
\item \href{#method-add_listener}{\code{Event$add_listener()}}
\item \href{#method-clone}{\code{Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialise an Event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name, }}{the name of the event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_listener"></a>}}
\subsection{Method \code{add_listener()}}{
Add an event listener
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$add_listener(listener)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listener, }}{the function to be executed on the event. This function
should take the simulation api and target individuals as arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
