% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{EventBase}
\alias{EventBase}
\title{EventBase Class}
\description{
Common functionality shared between simple and targeted events.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventBase-add_listener}{\code{EventBase$add_listener()}}
\item \href{#method-EventBase-.timestep}{\code{EventBase$.timestep()}}
\item \href{#method-EventBase-.tick}{\code{EventBase$.tick()}}
\item \href{#method-EventBase-.process}{\code{EventBase$.process()}}
\item \href{#method-EventBase-clone}{\code{EventBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBase-add_listener"></a>}}
\if{latex}{\out{\hypertarget{method-EventBase-add_listener}{}}}
\subsection{Method \code{add_listener()}}{
Add an event listener.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBase$add_listener(listener)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{listener}}{the function to be executed on the event, which takes a single
argument giving the time step when this event is triggered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBase-.timestep"></a>}}
\if{latex}{\out{\hypertarget{method-EventBase-.timestep}{}}}
\subsection{Method \code{.timestep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBase$.timestep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBase-.tick"></a>}}
\if{latex}{\out{\hypertarget{method-EventBase-.tick}{}}}
\subsection{Method \code{.tick()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBase$.tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBase-.process"></a>}}
\if{latex}{\out{\hypertarget{method-EventBase-.process}{}}}
\subsection{Method \code{.process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBase$.process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
