% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{execute_any_process}
\alias{execute_any_process}
\title{Execute a cpp or R process in the simulation}
\usage{
execute_any_process(p, api, cpp_api)
}
\arguments{
\item{p}{the process to execute}

\item{api}{the R api}

\item{cpp_api}{the C++ api}
}
\description{
Execute a cpp or R process in the simulation
}
