% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitset.R
\name{filter_bitset}
\alias{filter_bitset}
\title{Filter a bitset}
\usage{
filter_bitset(bitset, other)
}
\arguments{
\item{bitset}{the bitset to filter}

\item{other}{the values of bitset to keep}
}
\description{
filter in the i-th values in the set (assuming the bitset values
are in ascending order), for all i in other
}
