% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixed_probability_forked_state_change_process}
\alias{fixed_probability_forked_state_change_process}
\title{create a process to transition individuals from one state to multiple destination states at a constant rate}
\usage{
fixed_probability_forked_state_change_process(
  individual,
  source_state,
  destination_states,
  rate,
  destination_probabilities
)
}
\arguments{
\item{individual}{the name of an individual}

\item{source_state}{the name of the source state}

\item{destination_states}{vector of names of destination states}

\item{rate}{the rate at which each individual leaves}

\item{destination_probabilities}{a vector of probabilities for destination states upon leaving the source state}
}
\description{
create a process to transition individuals from one state to multiple destination states at a constant rate
}
