% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{fixed_probability_multinomial_process}
\alias{fixed_probability_multinomial_process}
\title{Multinomial process}
\usage{
fixed_probability_multinomial_process(
  variable,
  source_state,
  destination_states,
  rate,
  destination_probabilities
)
}
\arguments{
\item{variable}{a \code{\link{CategoricalVariable}} object}

\item{source_state}{a string representing the source state}

\item{destination_states}{a vector of strings representing the destination states}

\item{rate}{probability of individuals in source state to leave}

\item{destination_probabilities}{probability vector of destination states}
}
\value{
a function which can be passed as a process to \code{\link{simulation_loop}}
}
\description{
Simulates a two-stage process where all individuals
in a given \code{source_state} sample whether to leave or not with probability
\code{rate}; those who leave go to one of the \code{destination_states} with
probabilities contained in the vector \code{destination_probabilities}.
}
