% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixed_probability_state_change_process}
\alias{fixed_probability_state_change_process}
\title{create a process to transition individuals between states at a constant rate}
\usage{
fixed_probability_state_change_process(individual, from_state, to_state, rate)
}
\arguments{
\item{individual}{the name of an individual}

\item{from_state}{the name of the source state}

\item{to_state}{the name of the target state}

\item{rate}{the rate at which state transitions occur}
}
\description{
create a process to transition individuals between states at a constant rate
}
