% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{multi_probability_bernoulli_process}
\alias{multi_probability_bernoulli_process}
\title{Overdispersed Bernoulli process}
\usage{
multi_probability_bernoulli_process(variable, from, to, rate_variable)
}
\arguments{
\item{variable}{a \code{\link{CategoricalVariable}} object}

\item{from}{a string representing the source state}

\item{to}{a string representing the destination state}

\item{rate_variable}{\code{\link{DoubleVariable}} giving individual probability of each individual in source state to leave}
}
\value{
a function which can be passed as a process to \code{\link{simulation_loop}}
}
\description{
Simulates a Bernoulli process where all individuals
in a given source state \code{from} sample whether or not 
to transition to destination state \code{to} with a
individual probability specified by the \code{\link{DoubleVariable}}
object \code{rate_variable}.
}
