% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{Network}
\alias{Network}
\title{A static network class}
\description{
This class maintains an individual level contact network, using a
\code{\link[network]{network}} object to store and access edges. It only
handles static networks but is able to quickly compute a vector of contacts,
the number of infectious contacts on each susceptible. This vector can be used
to compute the force of infection on those susceptible individuals. It is able
to store both directed and undirected networks.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.contacts}}{a pointer to the underlying IntegerVariable}

\item{\code{.network}}{a \code{\link[network]{network}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Network$new()}}
\item \href{#method-compute_contacts}{\code{Network$compute_contacts()}}
\item \href{#method-get_contacts}{\code{Network$get_contacts()}}
\item \href{#method-clone}{\code{Network$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$new(g)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{g}}{a \code{\link[network]{network}} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-compute_contacts}{}}}
\subsection{Method \code{compute_contacts()}}{
insert into the bitset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$compute_contacts(S, I)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{a \code{\link{Bitset}} of susceptible individuals}

\item{\code{I}}{a \code{\link{Bitset}} of infectious individuals}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-get_contacts}{}}}
\subsection{Method \code{get_contacts()}}{
insert into the bitset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$get_contacts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
