% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_dataclass}
\alias{print_dataclass}
\title{Creates a generic print function for a DataClass}
\usage{
print_dataclass(class, fields)
}
\arguments{
\item{class}{the data classname}

\item{fields}{fields to print}
}
\description{
Creates a generic print function for a DataClass
}
