% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readonly_accessor}
\alias{readonly_accessor}
\title{Creates a readonly accessor for a private field of an R6 class}
\usage{
readonly_accessor(active, field)
}
\arguments{
\item{active}{the name of the accessor you want to create}

\item{field}{the field to create the accessor for}
}
\description{
Creates a readonly accessor for a private field of an R6 class
}
