% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{Render}
\alias{Render}
\title{Class: Render
Class to render output for the simulation}
\description{
Class: Render
Class to render output for the simulation

Class: Render
Class to render output for the simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Render$new()}}
\item \href{#method-add}{\code{Render$add()}}
\item \href{#method-to_dataframe}{\code{Render$to_dataframe()}}
\item \href{#method-clone}{\code{Render$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a renderer for the simulation, creates the default state
renderers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$new(timesteps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timesteps}}{number of timesteps in the simulation}

\item{\code{individuals}}{to render states for}

\item{\code{renderers}}{additional renderers to execute. Renderers are functions
which take a SimFrame as an argument and return a list of
scalar outputs to store to the final render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Update the render with new simulation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$add(name, value, timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the variable to render}

\item{\code{value}}{the value to store for the variable}

\item{\code{timestep}}{the timestep of the data point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Make a dataframe for the render
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$to_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
