% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reschedule_listener}
\alias{reschedule_listener}
\title{create a listener to schedule a target population for a new event}
\usage{
reschedule_listener(event, delay)
}
\arguments{
\item{event}{the name of the event to schedule}

\item{delay}{the delay for the new event}
}
\description{
create a listener to schedule a target population for a new event
}
