% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{save_object_state}
\alias{save_object_state}
\title{Save the state of a simulation object or set of objects.}
\usage{
save_object_state(objects)
}
\arguments{
\item{objects}{a simulation object (ie. a variable or event), or list
thereof.}
}
\value{
the saved states of the objects
}
\description{
Save the state of a simulation object or set of objects.
}
