% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{save_simulation_state}
\alias{save_simulation_state}
\title{Save the simulation state}
\usage{
save_simulation_state(timesteps, variables, events)
}
\arguments{
\item{timesteps}{the number of time steps that have already been simulated}

\item{variables}{the list of Variables}

\item{events}{the list of Events}
}
\value{
the saved simulation state.
}
\description{
Save the simulation state in an R object, allowing it to be
resumed later using \code{\link[individual]{restore_simulation_state}}.
}
