% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{Scheduler}
\alias{Scheduler}
\title{Class: Scheduler
Class to schedule events for the simulation}
\description{
Class: Scheduler
Class to schedule events for the simulation

Class: Scheduler
Class to schedule events for the simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-schedule}{\code{Scheduler$schedule()}}
\item \href{#method-get_scheduled}{\code{Scheduler$get_scheduled()}}
\item \href{#method-clear_schedule}{\code{Scheduler$clear_schedule()}}
\item \href{#method-process_events}{\code{Scheduler$process_events()}}
\item \href{#method-tick}{\code{Scheduler$tick()}}
\item \href{#method-get_timestep}{\code{Scheduler$get_timestep()}}
\item \href{#method-new}{\code{Scheduler$new()}}
\item \href{#method-clone}{\code{Scheduler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schedule"></a>}}
\subsection{Method \code{schedule()}}{
Schedule an event to occur in the future
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$schedule(event, target, delay)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event, }}{the event to schedule}

\item{\code{target, }}{the individuals to pass to the listener}

\item{\code{delay, }}{the number of timesteps to wait before triggering the event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scheduled"></a>}}
\subsection{Method \code{get_scheduled()}}{
Get the individuals who are scheduled for a particular event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$get_scheduled(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event, }}{the event of interest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_schedule"></a>}}
\subsection{Method \code{clear_schedule()}}{
Stop a future event from triggering for a subset of individuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$clear_schedule(event, target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event, }}{the event to stop}

\item{\code{target, }}{the individuals to clear}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_events"></a>}}
\subsection{Method \code{process_events()}}{
Get scheduled updates for the current timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$process_events(api, cpp_api)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api, }}{the api to pass to the listeners}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tick"></a>}}
\subsection{Method \code{tick()}}{
increment the scheduler clock
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_timestep"></a>}}
\subsection{Method \code{get_timestep()}}{
get current timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$get_timestep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
initialise the scheduler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$new(events, end_timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{events, }}{a list of event objects for the simulation}

\item{\code{end_timestep, }}{the number of timesteps to initialise for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
