% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate}
\alias{simulate}
\title{Main simulation loop}
\usage{
simulate(individuals, processes, end_timestep, parameters = list())
}
\arguments{
\item{individuals}{a list of Individual to simulate}

\item{processes}{a list of processes to execute on each timestep}

\item{end_timestep}{the number of timesteps to simulate}

\item{parameters}{a list of named parameters to pass to the process functions}
}
\description{
Main simulation loop
}
