% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Simulation}
\alias{Simulation}
\title{Class: Simulation
Class to store and update the simulation for each type of individual}
\description{
Class: Simulation
Class to store and update the simulation for each type of individual

Class: Simulation
Class to store and update the simulation for each type of individual
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render}{\code{Simulation$render()}}
\item \href{#method-get_api}{\code{Simulation$get_api()}}
\item \href{#method-tick}{\code{Simulation$tick()}}
\item \href{#method-apply_updates}{\code{Simulation$apply_updates()}}
\item \href{#method-new}{\code{Simulation$new()}}
\item \href{#method-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\subsection{Method \code{render()}}{
Return a list of the simulated states and variables for the simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$render(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the individual to render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_api"></a>}}
\subsection{Method \code{get_api()}}{
Get a SimFrame for the current timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$get_api()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tick"></a>}}
\subsection{Method \code{tick()}}{
Increment the timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_updates"></a>}}
\subsection{Method \code{apply_updates()}}{
Perform updates on the a simulation, increment the counter and return the
next simulation frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$apply_updates(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the updates as a list of update objects to apply}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a blank simulation and then initialize first timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a list of Individual and the number of timesteps to
initialise for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
