% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{State}
\alias{State}
\title{Class: State
Represents a state for an individual in our simulation}
\description{
Class: State
Represents a state for an individual in our simulation

Class: State
Represents a state for an individual in our simulation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{the string label for this state}

\item{\code{initial_size}}{the string label for this state}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{State$new()}}
\item \href{#method-clone}{\code{State$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new State
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$new(name, initial_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{is a unique idetifier which is used in the output}

\item{\code{initial_size}}{used to initialise the state at the start of the sim}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
