% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{state_count_renderer_process}
\alias{state_count_renderer_process}
\title{create a process to render the number of individuals in the specified states}
\usage{
state_count_renderer_process(individual, states)
}
\arguments{
\item{individual}{the name of an individual}

\item{states}{a vector of state names}
}
\description{
create a process to render the number of individuals in the specified states
}
