% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_private}
\alias{to_private}
\title{Create a private fieldname from an accessor name}
\usage{
to_private(name)
}
\arguments{
\item{name}{to convert}
}
\description{
Create a private fieldname from an accessor name
}
