% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{update_category_listener}
\alias{update_category_listener}
\title{Update category listener}
\usage{
update_category_listener(variable, to)
}
\arguments{
\item{variable}{a categorical variable}

\item{to}{a string representing the destination category}
}
\description{
Updates the category of a subpopulation as the result of an
event being triggered
}
