% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_state_listener}
\alias{update_state_listener}
\title{create a listener to update the state of the target population}
\usage{
update_state_listener(individual, state)
}
\arguments{
\item{individual}{the name of the individual type}

\item{state}{the state to transition to}
}
\description{
create a listener to update the state of the target population
}
