% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{Variable}
\alias{Variable}
\title{Class: Variable
Represents a variable for an individual in our simulation}
\description{
Class: Variable
Represents a variable for an individual in our simulation

Class: Variable
Represents a variable for an individual in our simulation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{the string label for this variable}

\item{\code{initialiser}}{the function to initialise the values for this variable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Variable$new()}}
\item \href{#method-clone}{\code{Variable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Variable. Variables represent a numerical value for each
individual. Variables are updated during a simulation when a process
returns a VariableUpdate object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$new(name, initialiser)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{is a unique identifier which is used in the output}

\item{\code{initialiser}}{a function used to initialise the variable at the start
of the simulation. The initialiser function takes the population size as
its only argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
