% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{variable_mean_renderer_process}
\alias{variable_mean_renderer_process}
\title{create a process to render the mean value of the specified variables}
\usage{
variable_mean_renderer_process(individual, variables)
}
\arguments{
\item{individual}{the name of an individual}

\item{variables}{a vector of variable names}
}
\description{
create a process to render the mean value of the specified variables
}
