% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{VariableUpdate}
\alias{VariableUpdate}
\title{Class: VariableUpdate
Describes an update to a variable}
\description{
Class: VariableUpdate
Describes an update to a variable

Class: VariableUpdate
Describes an update to a variable
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariableUpdate$new()}}
\item \href{#method-print}{\code{VariableUpdate$print()}}
\item \href{#method-clone}{\code{VariableUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new VariableUpdate descriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableUpdate$new(individual, variable, value, index = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{is the type of individual to update}

\item{\code{variable}}{a Variable object representing the variable to change}

\item{\code{value}}{a vector or scalar of values to assign at the index}

\item{\code{index}}{is the index at which to apply the change}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableUpdate$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
