% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitset.R
\docType{data}
\name{Bitset}
\alias{Bitset}
\title{A Bitset Class}
\format{
An object of class \code{list} of length 1.
}
\usage{
Bitset
}
\arguments{
\item{size}{the size of the bitset.}

\item{from}{pointer to an existing IterableBitset to use; if \code{NULL}
make empty bitset, otherwise copy existing bitset.}

\item{v}{an integer vector of elements (not indices) to remove.}

\item{inplace}{whether to overwrite the current bitset, default = TRUE}

\item{other}{the other bitset.}

\item{rate}{the success probability for keeping each element, can be
a single value for all elements or a vector of unique
probabilities for keeping each element.}

\item{k}{the number of items in the bitset to keep. The selection of
these k items from N total items in the bitset is random, and
k should be chosen such that \eqn{0 \le k \le N}.}
}
\description{
This is a data structure that compactly stores the presence of
integers in some finite set (\code{max_size}), and can
efficiently perform set operations (union, intersection, complement, symmetric
difference, set difference).
WARNING: All operations are in-place so please use \code{$copy}
if you would like to perform an operation without destroying your current bitset.

This class is defined as a named list for performance reasons, but for most
intents and purposes it behaves just like an R6 class.

create a bitset.

insert into the bitset.

remove from the bitset.

clear the bitset.

get the number of elements in the set.

to "bitwise or" or union two bitsets.

to "bitwise and" or intersect two bitsets.

to "bitwise not" or complement a bitset.

to "bitwise xor" or get the symmetric difference of two bitset
(keep elements in either bitset but not in their intersection).

Take the set difference of this bitset with another
(keep elements of this bitset which are not in \code{other}).

sample a bitset.

choose k random items in the bitset

returns a copy the bitset.

return an integer vector of the elements
stored in this bitset.
}
\keyword{datasets}
