% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{categorical_count_renderer_process}
\alias{categorical_count_renderer_process}
\title{Render Categories}
\usage{
categorical_count_renderer_process(renderer, variable, categories)
}
\arguments{
\item{renderer}{a \code{\link[individual]{Render}} object.}

\item{variable}{a \code{\link[individual]{CategoricalVariable}} object.}

\item{categories}{a character vector of categories to render.}
}
\value{
a function which can be passed as a process to \code{\link{simulation_loop}}.
}
\description{
Renders the number of individuals in each category.
}
