% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitset.R
\name{filter_bitset}
\alias{filter_bitset}
\title{Filter a bitset}
\usage{
filter_bitset(bitset, other)
}
\arguments{
\item{bitset}{the \code{\link{Bitset}} to filter}

\item{other}{the values to keep (may be a vector of intergers or another \code{\link{Bitset}})}
}
\description{
This non-modifying function returns a new \code{\link{Bitset}}
object of the same maximum size as the original but which only contains
those values at the indices specified by the argument \code{other}.
Indices in \code{other} may be specified either as a vector of integers or as
another bitset. Please note that filtering by another bitset is not a
"bitwise and" intersection, and will have the same behavior as providing
an equivalent vector of integer indices.
}
