% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{Individual}
\alias{Individual}
\title{Class: Individual
Represents an individual in our simulation}
\description{
Class: Individual
Represents an individual in our simulation

Class: Individual
Represents an individual in our simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Individual$new()}}
\item \href{#method-check_state}{\code{Individual$check_state()}}
\item \href{#method-check_variable}{\code{Individual$check_variable()}}
\item \href{#method-check_constant}{\code{Individual$check_constant()}}
\item \href{#method-print}{\code{Individual$print()}}
\item \href{#method-clone}{\code{Individual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Individual
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$new(name, states, variables = list(), constants = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{is a unique idetifier which is used in the output
$param ... a list of State objects
$param variables a list of Variable objects
$param constants a list of Constant objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_state"></a>}}
\subsection{Method \code{check_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$check_state(state)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_variable"></a>}}
\subsection{Method \code{check_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$check_variable(variable)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_constant"></a>}}
\subsection{Method \code{check_constant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$check_constant(constant)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
