% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{Individual}
\alias{Individual}
\title{Class: Individual
Represents an individual in our simulation}
\description{
Class: Individual
Represents an individual in our simulation

Class: Individual
Represents an individual in our simulation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name, }}{the string label for this individual}

\item{\code{states, }}{a list of state objects which apply to this individual}

\item{\code{variables, }}{a list of variable objects which apply to this individual}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name, }}{the string label for this individual}

\item{\code{states, }}{a list of state objects which apply to this individual}

\item{\code{variables, }}{a list of variable objects which apply to this individual}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Individual$new()}}
\item \href{#method-clone}{\code{Individual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Individual
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$new(name, states, variables = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{is a unique idetifier which is used in the output
$param states a list of State objects
$param variables a list of Variable objects
$param constants a list of Constant objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
