% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Render}
\alias{Render}
\title{Class: Render
Class to render output for the simulation}
\description{
Class: Render
Class to render output for the simulation

Class: Render
Class to render output for the simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Render$new()}}
\item \href{#method-update}{\code{Render$update()}}
\item \href{#method-to_dataframe}{\code{Render$to_dataframe()}}
\item \href{#method-clone}{\code{Render$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialise a renderer for the simulation, creates the default state
renderers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$new(individuals, timesteps, renderers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individuals}}{to render states for}

\item{\code{timesteps}}{number of timesteps in the simulation}

\item{\code{renderers}}{additional renderers to execute. Renderers are functions
which take a SimFrame as an argument and return a list of
scalar outputs to store to the final render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\subsection{Method \code{update()}}{
Update the render with new simulation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$update(frame, timestep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{frame}}{the new SimFrame}

\item{\code{timestep}}{the timestep of the frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\subsection{Method \code{to_dataframe()}}{
Make a dataframe for the render
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$to_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Render$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
