% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefab.R
\name{reschedule_listener}
\alias{reschedule_listener}
\title{Reschedule listener}
\usage{
reschedule_listener(event, delay)
}
\arguments{
\item{event}{a \code{\link[individual]{TargetedEvent}}.}

\item{delay}{the delay until the follow-up event.}
}
\description{
Schedules a follow-up event as the result of an event
firing.
}
