% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{restore_object_state}
\alias{restore_object_state}
\title{Restore the state of simulation objects.}
\usage{
restore_object_state(timesteps, objects, state)
}
\arguments{
\item{timesteps}{the number of time steps that have already been simulated}

\item{objects}{a simulation object (eg. a variable or event) or an
arbitrarily nested list structure of such objects.}

\item{state}{a saved simulation state for the given objects, as returned by
\code{\link[individual]{save_object_state}}. This should have the same shape
as the \code{objects} argument: if a list of objects is given, then
\code{state} should be a list of corresponding states. If NULL is passed,
then each object's \code{restore_state} method is called with NULL as
its argument.}
}
\description{
Restore the state of one or more simulation objects. The
specified objects are paired up with the relevant part of the state object,
and the \code{restore_state} method of each object is called.

If the list of object is named, more objects may be specified than were
originally present in the saved simulation, allowing a simulation to be
extended with more features upon resuming. In this case, the
\code{restore_state} method is called with a \code{NULL} argument.
}
