% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{restore_state}
\alias{restore_state}
\title{Restore the simulation state}
\usage{
restore_state(state, variables, events, restore_random_state)
}
\arguments{
\item{state}{the simulation state to restore, as returned by \code{\link[individual]{restore_state}}.}

\item{variables}{the list of Variables}

\item{events}{the list of Events}

\item{restore_random_state}{if TRUE, restore R's global random number generator's state from the checkpoint.}
}
\description{
Restore the simulation state from a previous checkpoint.
The state of passed events and variables is overwritten to match the state they
had when the simulation was checkpointed. Returns the time step at which the
simulation should resume.
}
