% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{save_object_state}
\alias{save_object_state}
\title{Save the state of a simulation object or set of objects.}
\usage{
save_object_state(objects)
}
\arguments{
\item{objects}{a simulation object (eg. a variable or event) or an
arbitrarily nested list structure of such objects.}
}
\value{
the saved states of the objects. This has the same shape as the given
\code{objects}: if a list was passed as an argument, this returns the
corresponding list of saved states. If a singular object was passed, this
returns just that particular object's state.
}
\description{
Save the state of a simulation object or set of objects.
}
