% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{SimAPI}
\alias{SimAPI}
\title{Class: SimAPI
The entry point for models to inspect and manipulate the simulation}
\description{
Class: SimAPI
The entry point for models to inspect and manipulate the simulation

Class: SimAPI
The entry point for models to inspect and manipulate the simulation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_state}{\code{SimAPI$get_state()}}
\item \href{#method-get_variable}{\code{SimAPI$get_variable()}}
\item \href{#method-queue_state_update}{\code{SimAPI$queue_state_update()}}
\item \href{#method-queue_variable_update}{\code{SimAPI$queue_variable_update()}}
\item \href{#method-schedule}{\code{SimAPI$schedule()}}
\item \href{#method-get_scheduled}{\code{SimAPI$get_scheduled()}}
\item \href{#method-clear_schedule}{\code{SimAPI$clear_schedule()}}
\item \href{#method-get_timestep}{\code{SimAPI$get_timestep()}}
\item \href{#method-get_parameters}{\code{SimAPI$get_parameters()}}
\item \href{#method-new}{\code{SimAPI$new()}}
\item \href{#method-clone}{\code{SimAPI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_state"></a>}}
\subsection{Method \code{get_state()}}{
Get the index of individuals with a particular state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$get_state(individual, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{of interest}

\item{\code{...}}{the states of interest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable"></a>}}
\subsection{Method \code{get_variable()}}{
Get a variable vector for an individual
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$get_variable(individual, variable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{the individual of interest}

\item{\code{variable}}{the variable of interest}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue_state_update"></a>}}
\subsection{Method \code{queue_state_update()}}{
Queue a state update for the end of the timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$queue_state_update(individual, state, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{the individual of interest}

\item{\code{state}}{the target state}

\item{\code{index}}{the index of individuals to move to the target state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue_variable_update"></a>}}
\subsection{Method \code{queue_variable_update()}}{
Queue a variable update for the end of the timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$queue_variable_update(individual, variable, index, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{the individual of interest}

\item{\code{variable}}{the variable to update}

\item{\code{index}}{the index of individuals to update}

\item{\code{values}}{the values to apply at index}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-schedule"></a>}}
\subsection{Method \code{schedule()}}{
Schedule an event to occur in the future
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$schedule(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{..., }}{forwarded to Scheduler$schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scheduled"></a>}}
\subsection{Method \code{get_scheduled()}}{
Get the individuals who are scheduled for a particular event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$get_scheduled(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{..., }}{forwarded to Scheduler$get_schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_schedule"></a>}}
\subsection{Method \code{clear_schedule()}}{
Stop a future event from triggering for a subset of individuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$clear_schedule(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{..., }}{forwarded to Scheduler$clear_schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_timestep"></a>}}
\subsection{Method \code{get_timestep()}}{
Get the current timestep of the simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$get_timestep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameters"></a>}}
\subsection{Method \code{get_parameters()}}{
Get the parameters of the simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$get_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create an R wrapper for the API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$new(cpp_api, scheduler, parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cpp_api, }}{the cpp implementation of the simulation api}

\item{\code{scheduler, }}{the implementation of the scheduler interface}

\item{\code{parameters, }}{model parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
