% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate}
\alias{simulate}
\title{Main simulation loop}
\usage{
simulate(
  individuals,
  processes,
  end_timestep,
  parameters = list(),
  events = list()
)
}
\arguments{
\item{individuals}{a list of Individual to simulate}

\item{processes}{a list of processes to execute on each timestep}

\item{end_timestep}{the number of timesteps to simulate}

\item{parameters}{a list of named parameters to pass to the process functions}

\item{events}{a list of events to register with the scheduler}
}
\description{
Main simulation loop
}
\examples{
population <- 4
S <- State$new('S', population)
I <- State$new('I', 0)
R <- State$new('R', 0)
human <- Individual$new('human', list(S, I, R))

transition <- function(from, to, rate) {
  return(function(api) {
    from_state <- api$get_state(human, from)
    api$queue_state_update(
      human,
      to,
      from_state[runif(length(from_state), 0, 1) < rate]
    )
  })
}

processes <- list(
  transition(S, I, .2),
  transition(I, R, .1),
  transition(R, S, .05)
)

simulate(human, processes, 5)
}
