% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{StateUpdate}
\alias{StateUpdate}
\title{Class: StateUpdate
Describes a state update}
\description{
Class: StateUpdate
Describes a state update

Class: StateUpdate
Describes a state update
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StateUpdate$new()}}
\item \href{#method-print}{\code{StateUpdate$print()}}
\item \href{#method-clone}{\code{StateUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new StateUpdate descriptor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateUpdate$new(individual, state, index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{individual}}{is the type of individual to update}

\item{\code{state}}{is the destination state of the update}

\item{\code{index}}{is the index at which to apply the change}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateUpdate$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
